+++ tests/helpers/helpers.js
@@ -2,6 +2,8 @@ import path from "path";
 import {DEFAULT_BRANCH, DIFF_SOURCE} from "../../src/constants";
 import fs from "fs";
 import yaml from "js-yaml";
+import {readYML} from "../../src/utils/common";
+import {diff_match_patch} from "diff-match-patch";
 
 export function getRandomString(length) {
     var randomChars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
@@ -125,3 +127,30 @@ export class Config {
         fs.writeFileSync(this.configPath, yaml.safeDump(config));
     }
 }
+
+export const assertChangeEvent = (repoPath, diffsRepo, oldText, updatedText,
+                                  fileRelPath, shadowFilePath,
+                                  diffsCount = 1) => {
+    // Read shadow file
+    const shadowText = fs.readFileSync(shadowFilePath, "utf8");
+    expect(shadowText).toStrictEqual(updatedText);
+    // Verify correct diff file has been generated
+    const diffFiles = fs.readdirSync(diffsRepo);
+    expect(diffFiles).toHaveLength(diffsCount);
+    const diffFilePath = path.join(diffsRepo, diffFiles[diffsCount-1]);
+    const diffData = readYML(diffFilePath);
+    expect(diffData.source).toEqual(DIFF_SOURCE);
+    expect(diffData.is_new_file).toBeFalsy();
+    expect(diffData.is_rename).toBeFalsy();
+    expect(diffData.is_deleted).toBeFalsy();
+    expect(diffData.repo_path).toEqual(repoPath);
+    expect(diffData.branch).toEqual(DEFAULT_BRANCH);
+    expect(diffData.file_relative_path).toEqual(fileRelPath);
+    // Verify diff is correct
+    const dmp = new diff_match_patch();
+    const patches = dmp.patch_make(oldText, updatedText);
+    //  Create text representation of patches objects
+    const diffs = dmp.patch_toText(patches);
+    expect(diffData.diff).toStrictEqual(diffs);
+    return true;
+};